refactor: dtos in chat folder
